Update endpoints for Cart router
Update endpoints for Cart router to only authorize actions on cartitems of the user

+++ routes/cartRouter.js
@@ -35,13 +35,8 @@ cartRouter
         res.end(`PUT operation not supported on /cartitems`);
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Cartitem.deleteMany()
-            .then(response => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(response);
-            })
-            .catch(err => next(err));
+        res.statusCode = 403;
+        res.end(`DELETE operation not supported on /cartitems`);
     });
 
 cartRouter
@@ -50,9 +45,15 @@ cartRouter
         Cartitem.findById(req.params.productId)
             .populate('user')
             .then(cartitem => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(cartitem);
+                if (cartitem.user.equals(req.user._id)) {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(cartitem);
+                } else {
+                    err = new Error('This is not an item in your cart!');
+                    err.status = 403;
+                    return next(err);
+                }
             })
             .catch(err => next(err));
     })
@@ -67,18 +68,34 @@ cartRouter
             { new: true }
         )
             .then(cartitem => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(cartitem);
+                if (cartitem.user.equals(req.user._id)) {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(cartitem);
+                } else {
+                    err = new Error('This is not an item in your cart!');
+                    err.status = 403;
+                    return next(err);
+                }
             })
             .catch(err => next(err));
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Cartitem.findByIdAndDelete(req.params.productId)
-            .then(response => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(response);
+        Cartitem.findById(req.params.productId)
+            .then(cartitem => {
+                if (cartitem.user.equals(req.user._id)) {
+                    Cartitem.findByIdAndDelete(req.params.productId)
+                        .then(response => {
+                            res.statusCode = 200;
+                            res.setHeader('Content-Type', 'application/json');
+                            res.json(response);
+                        })
+                        .catch(err => next(err));
+                } else {
+                    err = new Error('This is not an item in your cart!');
+                    err.status = 403;
+                    return next(err);
+                }
             })
             .catch(err => next(err));
     });
